build: Add basic Makefiles & main.c
Hello IFJ :wave:!

This surely works only for a single main.c file but I'm not sure about
it being very future resilient. The final binary + .o files can be found
in directory 'build'.

If a new source/header file is added, then it has to be added to
src/Makefile.

Most likely src/Makefile will be changed because I'd like to structure
the parts of a compiler into separate static parts that are then linked
together. We'll see.