token: Add header file for tokens
This is very basic. There's an enumerator with all the different types
of values the token can hold info about and the token struct itself. It
holds info about the token type and an arbitrary "attribute" that could
hold anything (e.g., value of int, nothing for operators, info about a
variable -> in this case it could be a struct with the name, if it is a
pointer, address where the pointer points).

For the more advanced attributes (as in the case of a string) there has
to be done some additional work (definition of the said struct,
functions for working with it,..). But this is still a subject for
future discussion.