scanner: Add basic scanner
This is a basic sketch of how the scanner could be implemented.
Currently it does not do anything useful.

It seems to me that the easiest way to go is to use finite state machine
as the heart of the scanner_scan function. I already noticed there's a
quite large overlap in the approach I'd like to take with tracking the
scanner's state. Most of the states are the same as token types in
token.h. I might reuse those in scanner.

Currently all member of the scanner_t struct are private. I wonder if
any will be needed to be public.