extras: Add hash table library from IJC
This library is the one I used (and wrote) in the second project in IJC
in 2019/2020. I did not include it as a submodule because I'm not
interested in maintaining it upstream and if any improvements will be
made, they'll be done here (and I'll mirror the code "upstream" but only
for showcase purposes).

The library is meant to be dynamically linked against, so invoking
"make" will cause a libhtab.so to be put in build/lib and header files
in build/include. This requires some options to be passed to the linker.
Those are handled using implicit variables LDFLAGS and LDLIBS[0].

In it's current state the hash-table cannot be really used with the
compiler because the items in a hash-table are not flexible enough
(currently a simple structure holding a string and an integer).

[0] https://www.gnu.org/software/make/manual/html_node/Implicit-Variables.html