src: Implement symtable
The first iteration of symtable. Builds on top of my IJC project from
previous semester (summer 2019/2020). A considerable number of changes
had to be made/I wanted to make:

- automatic reallocation of hashtable upon reaching a set limit
- change of type for data in entries
- tracking of on which element in a bucket the iterator is
- more advanced work with the data in entries
- minor fixes of original issues across the codebase

This is not final.

There's no way to destruct a value in a record because the definition
of tokens was not finalized at the time of writing.

There is no abstraction regarding stacks of symtables.

So... this is currently just a hashtable with fancy words slapped on top
of it :).