src/symtable: Add symtable manager
Symtable manager takes care of maintaining an array of symtables on
which it operates. It implements a set of functions that are similar to
the ones in symtable but with the difference that they work across all
the symtables in the "stack".

The set of functions for symtable manager is quite basic and I wouldn't
call them very efficient. My goal was to provide a basic set of
functions that can "get the job done" and when the needs of the compiler
gain a more solid form the symtable can be altered.

I tried to hunt down as many leaks and segfaults as possible but I'm not
100% that there are none left.