src: Add debug macros
Debug macros to make development a bit easier and not have to type
printf everywhere. To remove the debug prints, compile the binary with
-DNDEBUG.

Use these functions because the messages are printed to stderr by
default and print out from which file they're called and on what line.

debug() should be used in every case you'd use printf :). debug_entry()
should be used at the beginning of function you deem important enough
to be logged. It marks the entrance of said functions.

I considered added debug_exit() and debug_return() macros but that
proved to be much of a challenge :).

Code taken from https://github.com/zedshaw/learn-c-the-hard-way-lectures
and majorily cut down because I don't think we need all of it.