src/main: Support text passed throguh stdin
There is no check if stdin is empty because I don't know of a method
that both does not block the stream and works in environment not
connected to a terminal (e.g., CI).

Also adjusted tests to cover this use case and removed a test for
running the compiler without any arguments.