src/scanner: Streamline innit_scan
innit_scan managed the state of the scanner only partially. While not
incorrect, I did not like the style. Now, innit_scan only returns two
values: success and error (and technically speaking it could return void
because all the necessary info can be read from the scanner's state),
and correctly sets the scanner.

Thanks to this, writing tests is easier :).