Reimplement symbol table
While I was trying to wire symtable into parser, I realized the design
I came up with a month ago is not very flexible and very cumbersome to
use. Redesign time!

The new API is structured in a very different way. Most of the API is
hidden to the user and there is just a limited set of functions to use
with the symbol table. Those functions should be more than enough to use
it.

Changelog:
- Most of the API were made private
- Many changes to all functions (deleted/added/modified)
- New level of abstraction - symtable_scope - hidden under the hood and
  only exposed to some extent - used for function scopes
- symtable_stack (formerly manager) is implemented as linked list
- symtable_t incorporates symtable_scope, symtable_stack and symtable_row