Implement Stack data structure holding any type
Stack is implemented as a dynamic array.

Because I wanted to make the stack as general as possible due to the
request from the codegen team, I used several obfuscating macros to
change the data type, stack struct and all its functions to compute
their name at the compile time when including the file.

Look at the usage in "stack.h" to better understand the magic. I don't
think the time spent making this work was worth it but it was
interesting for sure :D

Reordered the included files in Makefile alphabetically.