ast: Add types for abstract syntactic tree
I came to the conclusion that I need some data structures where to store
info about entities like functions (mainly and currently solely).

I was in a rush and simply made the types in a way that they should be
allocated on the heap. Optimization should come but I doubt it will :).