Change charptr type to match the strcmp implementation
Switch from char* to const char* because of how the 'strcmp' function
code generation team uses is implemented. There is no functional
difference between char* and const char* regarding the stack
implementation and its functionality, so this is safe to do. (Once the
char* is pushed, it is no longer tempered with.)